backend/requirements.txt
Flask==2.3.3
flask-cors==3.1.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
PyJWT==2.8.0
gevent==22.10.2          # untuk WebSocket
Flask-Sockets==0.2.1
python-dotenv==1.0.1
